Implement the first iteration of components (#3)
This patch decouples the request form, the template generation and
the preview.

Markup.js is used for better templatingi, with default variables and
minimal logic. This allows us to populate only the sections
containing data provided by the user.

The preview is rendered by generating HTML from the markdown
template.