Add a component to display suggestions
This patch adds a new component that displays a list of business
suggestions based on GPS coordinates.

When retrieving information from Google maps, a list of places located
near by is returned. Depending on the accuracy of the coordinates or
whether there are multiple businesses in the same building, we cannot
accurately determine which business the user wanted to specify.
Therefore the list of available possibilities is displayed in a
`vaadin-grid` component and the user can select the most appropriate
one.