Improve Dockerfile
Creates a multi-stage `Dockerfile` to build an ES5 version of the
project and then serving it using Nginx.

Also creates a `Dockerfile.dev` to serve the project using the
`polymer-cli` with the live-reload option. This should be useful in the
feature to develop locally and test the project in real time on
Minikube

Drive-by:
* Fix the service ports.