Fix docker image build (#72)
The Docker image currently does not build because the `config.js` file
is absent. This patch inserts public information that are valid for the
production environment into the project.

The tools were updated to generate developer configurations.

The `Dockerfile.dev` file must be used to build an image that is
suitable to use in minikube.